backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.12
qdrant-client==1.16.1
google-generativeai==0.4.0
python-dotenv==1.0.0
httpx==0.24.1

